✨ [Feat] : Request validation 기능 추가, 데이터베이스 CRUD API 기능 추가, TaskError 타입 추가함

+++ ProjectManagerServer/Sources/App/Migrations/TaskMigration.swift
@@ -13,7 +13,7 @@ struct TaskMigration: Migration {
                 .id()
                 .field("title", .string)
                 .field("description", .string)
-                .field("date", .double)
+                .field("date", .int)
                 .field("status", status)
                 .create()
         }
